working

+++ P3/1401_Fernandez_Gonzalez-Carvajal/sql_files/getTopMonths.sql
@@ -3,7 +3,7 @@ CREATE OR REPLACE FUNCTION getTopMonths (umbralprods integer, umbralimport numer
     year double precision,
     month double precision,
     price numeric,
-    products bigint
+    products numeric
 ) as $$
 BEGIN
   RETURN QUERY
@@ -14,11 +14,18 @@ BEGIN
       DATE_PART('month', orders.orderdate) as month, 
       orders.totalamount, 
       SUM(orderdetail.quantity) as quantity
-  FROM orders INNER JOIN orderdetail ON (orders.orderid=orderdetail.orderid)
+  FROM orders
+    INNER JOIN orderdetail ON (
+      orders.orderid=orderdetail.orderid)
   GROUP BY orders.orderid, year, month, totalamount
   )
-  SELECT tablaaux.year, tablaaux.month, SUM(tablaaux.totalamount), SUM(tablaaux.quantity)
-  FROM tablaaux
+  SELECT
+    tablaaux.year,
+    tablaaux.month,
+    SUM(tablaaux.totalamount),
+    SUM(tablaaux.quantity)
+  FROM
+    tablaaux
   GROUP BY tablaaux.year, tablaaux.month
   HAVING SUM(tablaaux.totalamount) >= umbralimport OR SUM(tablaaux.quantity) >= umbralprods
   ORDER BY tablaaux.year, tablaaux.month;
